+++ frontend/src/App.tsx
@@ -4,20 +4,20 @@ import { Suspense, lazy } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { RecoilRoot } from 'recoil';
 
-import GptResultModal from './components/GptResultModal';
-import LoadingSpinner from './components/LoadingSpinner';
+import GptResultModal from './components/general/GptResultModal';
+import LoadingSpinner from './components/general/LoadingSpinner';
 
 // lazy 동적으로 필요할 때 import를 하여 실제로 로드되는 것
-const MainPage = lazy(() => import('./pages/MainPage'));
-const MyPage = lazy(() => import('./pages/MyPage'));
-const WritingPage = lazy(() => import('./pages/WritingPage'));
-const ReadingPage = lazy(() => import('./pages/ReadingPage'));
-const ModifyingPage = lazy(() => import('./pages/ModifyingPage'));
-const ErrorPage = lazy(() => import('./pages/ErrorPage'));
-const SigninPage = lazy(() => import('./pages/SigninPage'));
-const SignupPage = lazy(() => import('./pages/SignupPage'));
-const SearchProjectPage = lazy(() => import('./pages/SearchProjectPage'));
-const SearchUserPage = lazy(() => import('./pages/SearchUserPage'));
+const MainPage = lazy(() => import('./pages/main/MainPage'));
+const MyPage = lazy(() => import('./pages/user/MyPage'));
+const WritingPage = lazy(() => import('./pages/submit/WritingPage'));
+const ReadingPage = lazy(() => import('./pages/read/ReadingPage'));
+const ModifyingPage = lazy(() => import('./pages/submit/ModifyingPage'));
+const ErrorPage = lazy(() => import('./pages/error/ErrorPage'));
+const SigninPage = lazy(() => import('./pages/user/SigninPage'));
+const SignupPage = lazy(() => import('./pages/user/SignupPage'));
+const SearchProjectPage = lazy(() => import('./pages/main/SearchProjectPage'));
+const SearchUserPage = lazy(() => import('./pages/main/SearchUserPage'));
 
 const queryClient = new QueryClient();
