+++ frontend/src/components/LikeModal.tsx
@@ -1,8 +1,5 @@
 import React from 'react';
 
-// import closeIcon from '../assets/image/closeIcon.svg';
-// import likeIcon from '../assets/image/likeIcon.svg';
-
 type LikeModalProps = {
   onClickLikeToggleModal: () => void;
   isLike: boolean;
@@ -17,35 +14,17 @@ function likeModal({ onClickLikeToggleModal, isLike }: LikeModalProps) {
     bg-white sm:w-630 sm:py-5"
       >
         {isLike}
-        {/* <div className="flex justify-center font-lato text-[23px] font-semibold">
-          <div className="text-graphyblue">반응한 사람</div>
-          <img className="" src={closeIcon} alt="closeIcon" />
-        </div> */}
-
-        {/* 구분선 */}
-        {/* <div className="my-3 w-auto border-b-2 border-b-neutral-200" />
-
-        <div className="my-6 flex items-center justify-items-start px-10 font-lato">
-          <img className="mr-4 w-14" src={likeIcon} alt="likeIcon" />
-
-          <div className="mr-2 whitespace-nowrap  text-[21px]  font-semibold text-graphyblue">
-            강민아
-          </div>
-          <div className=" text-left text-[17px] font-normal text-zinc-500">
-            본인을 소개하는 한 마디
-          </div>
-        </div> */}
       </div>
 
-      {/* 모달 영역 외의 배경을 클릭하면 모달이 닫히게 하는 컨테이너, 이벤트핸들러를 사용하여 클릭 이벤트 발생 시 onClickToggleModal 함수 호출하여 모달 닫음 */}
+      {/* 모달 배경 */}
       <button
         aria-label="Toggle modal"
         className="fixed top-0 left-0 right-0 bottom-0 z-40 h-full w-screen  bg-black opacity-70"
         onClick={(e: React.MouseEvent) => {
           e.preventDefault();
 
-          if (onClickToggleModal) {
-            onClickToggleModal();
+          if (onClickLikeToggleModal) {
+            onClickLikeToggleModal();
           }
         }}
         type="button"
