+++ frontend/src/components/main/Banner.tsx
@@ -4,7 +4,7 @@ import { Swiper, SwiperSlide } from 'swiper/react';
 import 'swiper/css';
 import 'swiper/scss/pagination';
 
-import lottie from '../assets/lottie';
+import lottie from '../../assets/lottie';
 
 function Banner() {
   return (
